# docs/swagger.json
{"schemes":["http"],"swagger":"2.0","info":{"description":"Transaction Ledger API for managing users and accounts","title":"Transaction Ledger API","contact":{},"version":"0.1.0"},"host":"localhost:8080","basePath":"/","paths":{"/accounts":{"post":{"description":"Creates a new account","produces":["application/json"],"summary":"Create a new account","parameters":[{"description":"Account creation payload","name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/api.AccountPayload"}}],"responses":{"201":{"description":"Account object","schema":{"$ref":"#/definitions/model.Account"}},"400":{"description":"Invalid request payload","schema":{"$ref":"#/definitions/api.ErrorResponse"}},"500":{"description":"Failed to create account","schema":{"$ref":"#/definitions/api.ErrorResponse"}}}}},"/accounts/{accountId}":{"get":{"description":"Retrieves an account by its ID","produces":["application/json"],"summary":"Get account","parameters":[{"type":"string","format":"uuid","description":"Account ID","name":"accountId","in":"path","required":true}],"responses":{"200":{"description":"Account object","schema":{"$ref":"#/definitions/model.Account"}},"400":{"description":"Invalid account ID","schema":{"$ref":"#/definitions/api.ErrorResponse"}},"404":{"description":"Account not found","schema":{"$ref":"#/definitions/api.ErrorResponse"}}}}},"/health":{"get":{"description":"Returns 200 OK if the API is running","produces":["text/plain"],"summary":"API Health check","responses":{"200":{"description":"OK","schema":{"type":"string"}}}}},"/transactions":{"post":{"description":"Creates a new transaction for an account","produces":["application/json"],"summary":"Create a new transaction","parameters":[{"description":"Transaction payload","name":"transaction","in":"body","required":true,"schema":{"$ref":"#/definitions/api.TransactionPayload"}}],"responses":{"201":{"description":"Transaction object","schema":{"$ref":"#/definitions/model.Transaction"}},"400":{"description":"Invalid request payload","schema":{"$ref":"#/definitions/api.ErrorResponse"}},"500":{"description":"Failed to create transaction","schema":{"$ref":"#/definitions/api.ErrorResponse"}}}}},"/transactions/{transactionId}":{"get":{"description":"Retrieves a transaction by its ID","produces":["application/json"],"summary":"Get transaction","parameters":[{"type":"string","format":"uuid","description":"Transaction ID","name":"transactionId","in":"path","required":true}],"responses":{"200":{"description":"Transaction object","schema":{"$ref":"#/definitions/model.Transaction"}},"400":{"description":"Invalid transaction ID","schema":{"$ref":"#/definitions/api.ErrorResponse"}},"404":{"description":"Transaction not found","schema":{"$ref":"#/definitions/api.ErrorResponse"}}}}},"/users":{"post":{"description":"Creates a new user in the system","produces":["application/json"],"summary":"Create a new user","responses":{"201":{"description":"User object","schema":{"$ref":"#/definitions/api.UserResponse"}},"500":{"description":"Failed to create user","schema":{"$ref":"#/definitions/api.ErrorResponse"}}}}},"/users/{id}":{"get":{"description":"Retrieves a user by their ID","produces":["text/plain"],"summary":"Get user by ID","parameters":[{"type":"string","format":"uuid","description":"User ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"User object","schema":{"$ref":"#/definitions/api.UserResponse"}},"400":{"description":"Invalid user ID","schema":{"$ref":"#/definitions/api.ErrorResponse"}},"404":{"description":"User not found","schema":{"$ref":"#/definitions/api.ErrorResponse"}}}}}},"definitions":{"api.AccountPayload":{"type":"object","required":["balance","userId"],"properties":{"balance":{"description":"Balance is the initial balance of the account\n\t@example\t1000","type":"integer"},"userId":{"description":"UserId is the unique identifier of the user who owns the account\n\t@example\t660e8400-e29b-41d4-a716-446655440000","type":"string"}}},"api.ErrorResponse":{"type":"object","properties":{"error":{"type":"string"}}},"api.TransactionPayload":{"type":"object","required":["accountId","amount","type"],"properties":{"accountId":{"description":"AccountId is the unique identifier of the account associated with the transaction\n\t@example\t880e8400-e29b-41d4-a716-446655440000","type":"string"},"amount":{"description":"Amount is the amount of the transaction\n\t@example\t500","type":"integer"},"type":{"description":"Type is the type of the transaction (e.g., credit or debit) as an integer\n\t@example\t0","type":"integer"}}},"api.UserResponse":{"type":"object","properties":{"createdAt":{"description":"CreatedAt is the timestamp when the user was created\n\t@example\t2025-12-25T11:11:00Z","type":"string"},"id":{"description":"Id is the unique identifier of the user\n\t@example\t550e8400-e29b-41d4-a716-446655440000","type":"string"}}},"model.Account":{"type":"object","properties":{"balance":{"type":"integer"},"createdAt":{"description":"Milliseconds since epoch","type":"integer"},"id":{"type":"string"},"userId":{"type":"string"}}},"model.Transaction":{"type":"object","properties":{"accountId":{"type":"string"},"amount":{"type":"integer"},"createdAt":{"description":"Milliseconds since epoch","type":"integer"},"id":{"type":"string"},"type":{"description":"Credit or Debit","type":"integer"}}}}}